+++ core/src/commonTest/kotlin/com/powersync/bucket/BucketStorageTest.kt
@@ -79,6 +79,7 @@ class BucketStorageTest {
                         mapOf(
                             "key" to "value",
                         ),
+                    data = mapOf("key" to "value"),
                 )
             mockDb =
                 mock<InternalDatabase> {